Update Migrate File

+++ db/migrate/20150227030543_add_column_user_group_id_and_remove_role.rb
@@ -1,6 +1,12 @@
 class AddColumnUserGroupIdAndRemoveRole < ActiveRecord::Migration
   def change
-    add_column :users, :user_group_id, :string, limit: 36, null: false
-    remove_column :users, :role
+  	if column_exists? :users, :role
+	  remove_column :users, :role
+	end
+  	
+  	unless column_exists? :users, :user_group_id
+		add_column :users, :user_group_id, :string, limit: 36, null: false
+	end
+    
   end
 end
